Update wavelet.py
Hello, Thanks for this great tool. Here is s small bug. When I use the significance function and set signal as an array-like series, the returned significance only includes a part of the red-noise spectrum. 
Based on Eq. 16 in Torrence & Compo (1998), k = 0 ... N/2 is the frequency index. 
In line 257, the inputed 'freq' = dt / period, which is the normalized frequency.
So, N should not be len(signal) anymore. N should be 1.